+++ lib/src/cli/git_cli.dart
@@ -1,8 +1,7 @@
 part of 'cli.dart';
 
 /// {@template unreachable_git_dependency}
-/// Thrown when `flutter packages get` or `flutter pub get`
-/// encounters an unreachable git dependency.
+/// Thrown when `flutter pub get` encounters an unreachable git dependency.
 /// {@endtemplate}
 class UnreachableGitDependency implements Exception {
   /// {@macro unreachable_git_dependency}