+++ lib/src/commands/create/templates/very_good_flutter_package/very_good_flutter_package_template.dart
@@ -17,8 +17,9 @@ class FlutterPkgTemplate extends Template {
 
   @override
   Future<void> onGenerateComplete(Logger logger, Directory outputDir) async {
-    await installFlutterPackages(logger, outputDir);
-    await applyDartFixes(logger, outputDir);
+    if (await installFlutterPackages(logger, outputDir)) {
+      await applyDartFixes(logger, outputDir);
+    }
     _logSummary(logger);
   }